touched up on some error messages and verification that we forgot about

+++ code/src/client/viewmodel/BranchViewModel.java
@@ -11,10 +11,14 @@ import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import shared.Branch.Branch;
 import shared.PropertyChangeSubject;
+import shared.Reservation.Car;
+import shared.Reservation.Reservation;
+import shared.personel.Employee;
 
 import java.beans.PropertyChangeEvent;
 import java.beans.PropertyChangeListener;
 import java.beans.PropertyChangeSupport;
+import java.lang.reflect.Array;
 import java.util.ArrayList;
 
 public class BranchViewModel {
@@ -22,13 +26,54 @@ public class BranchViewModel {
     private final ViewHandler viewHandler;
     private ObservableList<Branch> branches;
 
+    private ArrayList<Car> cars;
+    private ArrayList<Reservation> reservations;
+    private ArrayList<Employee> employees;
+
+    private StringProperty message;
+
     public BranchViewModel(Model model, ViewHandler viewHandler) {
-        branches = FXCollections.observableArrayList();
+
         this.model = model;
+        this.viewHandler = viewHandler;
+
+        branches = FXCollections.observableArrayList();
+        cars = new ArrayList<>();
+        reservations = new ArrayList<>();
+        employees = new ArrayList<>();
+
+        model.addListener(this::listenForBranches, "branches");
+        model.addListener(this::listenForCars, "cars");
+        model.addListener(this::listenForEmployees, "employees");
+        model.addListener(this::listenForReservations, "reservations");
 
         model.getBranches();
-        this.viewHandler = viewHandler;
-        model.addListener(this::listenForBranches, "branches");
+
+
+        message = new SimpleStringProperty();
+
+    }
+
+    public StringProperty messageProperty () {
+        return message;
+    }
+
+    private void listenForReservations(PropertyChangeEvent propertyChangeEvent) {
+        Platform.runLater(() -> {
+            reservations = (ArrayList<Reservation>) propertyChangeEvent.getNewValue();
+        });
+    }
+
+    private void listenForEmployees(PropertyChangeEvent propertyChangeEvent) {
+        Platform.runLater(() -> {
+            employees = (ArrayList<Employee>) propertyChangeEvent.getNewValue();
+        });
+    }
+
+    private void listenForCars(PropertyChangeEvent propertyChangeEvent) {
+        Platform.runLater(() -> {
+            cars = (ArrayList<Car>) propertyChangeEvent.getNewValue();
+        });
     }
 
 
@@ -64,8 +109,48 @@ public class BranchViewModel {
     }
 
     public void deleteAction (int id) {
-        model.deleteBranch(id);
-        model.getBranches();
+        //TODO associate message property with a message in the GUI, i didnt do it right now because i didnt want to touch the views so there are no conflicts - OLIVER
+        model.getReservations();
+        model.getCars();
+        model.getEmployees();
+        boolean canDelete = true;
+        boolean employeeAssociation = false;
+        boolean carAssociation = false;
+        boolean reservationAssociation = false;
+
+
+        for (Employee employee : employees) {
+            if (employee.getBranchId() == id) {
+                employeeAssociation = true;
+                canDelete = false;
+            }
+        }
+
+        for (Car car : cars) {
+            if (car.getBranchId() == id) {
+                carAssociation = true;
+                canDelete = false;
+            }
+        }
+
+        for (Reservation reservation : reservations) {
+            if (reservation.getStartBranchId() == id || reservation.getEndBranchId() == id){
+                reservationAssociation = true;
+                canDelete = false;
+            }
+        }
+
+        if (canDelete) {
+            model.deleteBranch(id);
+            model.getBranches();
+        } else {
+            if (employeeAssociation)
+                message.set("This branch cannot be deleted because there are employees associated with it");
+            else if (carAssociation)
+                message.set("This branch cannot be deleted because there are cars associated with it");
+            else if (reservationAssociation)
+                message.set("This branch cannot be deleted because there are reservations associated with it");
+        }
     }
 
 }
