+++ code/src/client/viewmodel/CarViewModel.java
@@ -3,9 +3,12 @@ package client.viewmodel;
 import client.core.ViewHandler;
 import client.model.Model;
 import javafx.application.Platform;
+import javafx.beans.property.SimpleStringProperty;
+import javafx.beans.property.StringProperty;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import shared.Reservation.Car;
+import shared.Reservation.Reservation;
 
 import java.beans.PropertyChangeEvent;
 import java.util.ArrayList;
@@ -15,13 +18,28 @@ public class CarViewModel {
     private Model model;
     private ViewHandler viewHandler;
     private ObservableList<Car> cars;
-
+    private ArrayList<Reservation> reservations;
+    private StringProperty message;
 
     public CarViewModel(Model model, ViewHandler viewHandler) {
         cars = FXCollections.observableArrayList();
+        reservations = new ArrayList<>();
         this.model = model;
         this.viewHandler = viewHandler;
+        message = new SimpleStringProperty();
         model.addListener(this::listenForCars, "cars");
+        model.addListener(this::listenForReservations, "reservations");
+        model.getCars();
+    }
+
+    public StringProperty messageProperty () {
+        return message;
+    }
+
+    private void listenForReservations(PropertyChangeEvent propertyChangeEvent) {
+        Platform.runLater(() -> {
+            reservations = (ArrayList<Reservation>) propertyChangeEvent.getNewValue();
+        });
     }
 
     public void listenForCars(PropertyChangeEvent event) {
@@ -38,12 +56,6 @@ public class CarViewModel {
         viewHandler.openFrontPageView();
     }
 
-    public void addCar() {
-
-        viewHandler.openAddCarView();
-        model.getBranches();
-    }
-
     public ObservableList<Car> getCars() {
         return cars;
     }
@@ -57,7 +69,22 @@ public class CarViewModel {
     }
 
     public void deleteAction (int id) {
-        model.deleteCar(id);
-        model.getCars();
+
+        //TODO associate message property with a message in the GUI, i didnt do it right now because i didnt want to touch the views so there are no conflicts - OLIVER
+
+        model.getReservations();
+
+        boolean canDelete = true;
+
+        for (Reservation reservation : reservations) {
+            if (reservation.getCarId() == id)
+                canDelete = false;
+        }
+        if (canDelete) {
+            model.deleteCar(id);
+            model.getCars();
+        } else {
+            message.set("This car cannot be deleted because there is a reservation associated with it");
+        }
     }
 }
